+++ test/test_gradient.jl
@@ -1,9 +1,11 @@
 using SparseConnectivityTracer
 using SparseConnectivityTracer: GradientTracer, Dual, MissingPrimalError, trace_input
+using Test
+
+using Random: rand, GLOBAL_RNG
 using LinearAlgebra: det, dot, logdet
 using SpecialFunctions: erf, beta
 using NNlib: NNlib
-using Test
 
 # Load definitions of GRADIENT_TRACERS, GRADIENT_PATTERNS, HESSIAN_TRACERS and HESSIAN_PATTERNS
 include("tracers_definitions.jl")
@@ -63,7 +65,6 @@ REAL_TYPES = (Float64, Int, Bool, UInt8, Float16, Rational{Int})
         @test J(x -> (2//3)^x, 1) ≈ [1;;]
         @test J(x -> x^ℯ, 1) ≈ [1;;]
         @test J(x -> ℯ^x, 1) ≈ [1;;]
-        @test J(x -> round(x, RoundNearestTiesUp), 1) ≈ [0;;]
         @test J(x -> 0, 1) ≈ [0;;]
 
         # Test special cases on empty tracer
@@ -71,6 +72,18 @@ REAL_TYPES = (Float64, Int, Bool, UInt8, Float16, Rational{Int})
         @test J(x -> (2//3)^zero(x), 1) ≈ [0;;]
         @test J(x -> zero(x)^ℯ, 1) ≈ [0;;]
         @test J(x -> ℯ^zero(x), 1) ≈ [0;;]
+
+        # Round
+        @test J(round, 1.1) ≈ [0;;]
+        @test J(x -> round(Int, x), 1.1) ≈ [0;;]
+        @test J(x -> round(Bool, x), 1.1) ≈ [0;;]
+        @test J(x -> round(Float16, x), 1.1) ≈ [0;;]
+        @test J(x -> round(x, RoundNearestTiesAway), 1.1) ≈ [0;;]
+        @test J(x -> round(x; digits=3, base=2), 1.1) ≈ [0;;]
+
+        # Random
+        @test J(x -> rand(typeof(x)), 1) ≈ [0;;]
+        @test J(x -> rand(GLOBAL_RNG, typeof(x)), 1) ≈ [0;;]
 
         # Linear Algebra
         @test J(x -> dot(x[1:2], x[4:5]), rand(5)) == [1 1 0 1 1]
@@ -202,8 +215,18 @@ end
         @test J(x -> (2//3)^x, 1) ≈ [1;;]
         @test J(x -> x^ℯ, 1) ≈ [1;;]
         @test J(x -> ℯ^x, 1) ≈ [1;;]
-        @test J(x -> round(x, RoundNearestTiesUp), 1) ≈ [0;;]
         @test J(x -> 0, 1) ≈ [0;;]
+
+        # Round
+        @test J(round, 1.1) ≈ [0;;]
+        @test J(x -> round(Int, x), 1.1) ≈ [0;;]
+        @test J(x -> round(Bool, x), 1.1) ≈ [0;;]
+        @test J(x -> round(x, RoundNearestTiesAway), 1.1) ≈ [0;;]
+        @test J(x -> round(x; digits=3, base=2), 1.1) ≈ [0;;]
+
+        # Random
+        @test J(x -> rand(typeof(x)), 1) ≈ [0;;]
+        @test J(x -> rand(GLOBAL_RNG, typeof(x)), 1) ≈ [0;;]
 
         # Linear algebra
         @test J(logdet, [1.0 -1.0; 2.0 2.0]) == [1 1 1 1]  # (#68)
@@ -213,6 +236,12 @@ end
         # NNlib extension
         @test J(NNlib.relu, -1) ≈ [0;;]
         @test J(NNlib.relu, 1) ≈ [1;;]
+        @test J(NNlib.elu, -1) ≈ [1;;]
+        @test J(NNlib.elu, 1) ≈ [1;;]
+        @test J(NNlib.celu, -1) ≈ [1;;]
+        @test J(NNlib.celu, 1) ≈ [1;;]
+        @test J(NNlib.selu, -1) ≈ [1;;]
+        @test J(NNlib.selu, 1) ≈ [1;;]
 
         @test J(NNlib.relu6, -1) ≈ [0;;]
         @test J(NNlib.relu6, 1) ≈ [1;;]
@@ -220,6 +249,14 @@ end
 
         @test J(NNlib.trelu, 0.9) ≈ [0;;]
         @test J(NNlib.trelu, 1.1) ≈ [1;;]
+
+        @test J(NNlib.swish, -5) ≈ [1;;]
+        @test J(NNlib.swish, 0) ≈ [1;;]
+        @test J(NNlib.swish, 5) ≈ [1;;]
+
+        @test J(NNlib.hardswish, -5) ≈ [0;;]
+        @test J(NNlib.hardswish, 0) ≈ [1;;]
+        @test J(NNlib.hardswish, 5) ≈ [1;;]
 
         @test J(NNlib.hardσ, -4) ≈ [0;;]
         @test J(NNlib.hardσ, 0) ≈ [1;;]